Rewrite multipleSeatHoldRequests as a parameterized test
This is actually a bit more complicated since it:

  1. Introduces a new dependency (junit-jupiter-params)
  2. Requires another, persistent TicketService to be set up (to track state across test executions, which could be troublesome)

The test itself is a bit simpler, which is good, but there is lots of additional, slightly hidden complexity, which is bad. I'm
not convinced that for this case it is an appropriate usage, however I'll leave this here for demonstration purposes.