Start to implement actual reservations
This changes from a queue to keep the SeatHolds to a LinkedHashMap. This keeps insertion order intact,
which allows us to emulate a queue for purposes of expiring SeatHolds, while retaining the ability to
retrieve individual holds by their ID directly for reservations.

Also adds a printSeats convenience method to be able to see the current state of the venue.