Add true best seat picking algorithm
Adds a best seat selection algorithm that selects based on distance from front,
center seat. It does this by assigning a (0,0) coordinate to the center and calculating
the distance on the XY plane from that point. Simply said: the closer to the origin, the
"better" the seat.

This changes the "Goodness" score from an int to a double, and also flips around the
meaning of the value from highest being best to lowest being best. Luckily it was early
enough in the process that we didn't have to deal with API users. ;]

Mostly for demonstration purposes I added a displaySeatReservationProgress test that
shows how the venue is filled (but does spam test output quite a bit).