Simplify seat goodness calculation
I was using  for the 'distance from front/center' calculation, but
the sqrt was unnecessary, so I removed that. Also had a negative sign in front of the
X position, which really doesn't matter in terms of which side gets assigned seats
first, so I removed that as well.

Also shortened how many seats the  test has to display.